refactor callguideindexclient
