+++ src/components/call-guide/colors.ts
@@ -1,0 +1,8 @@
+export const partColors = {
+  MIKU: '#39c5bbaa',
+  RIN: '#ffa500aa',
+  LEN: '#ffe211aa',
+  LUKA: '#ffc0cbaa',
+  KAITO: '#0000ffaa',
+  MEIKO: '#d80000aa',
+} as const;
