+++ src/types/callGuide.ts
@@ -1,0 +1,17 @@
+import type { Prisma } from '@prisma/client';
+
+export type SongWithSetlist = Prisma.SongGetPayload<{
+  include: {
+    setlists: {
+      select: { order: true; higawari: true; locationgawari: true };
+      orderBy: { order: 'asc' };
+      take: 1;
+    };
+  };
+}>;
+
+export interface Playlist {
+  name: string;
+  slugs: string[];
+  color?: string;
+}
